refactor and code cleanup
