+++ index.html
@@ -26,9 +26,9 @@
       eval(variableName + " = object")
     }
   </script>
-  <script type="py" src="/src/main.py"></script>
+  <script type="py" src="/src/pyscript/main.py"></script>
   <script type="module" src="/src/main.ts"></script>
-  <script type="py" src="/src/worker.py" worker config="/src/pyconfig.json"></script>
+  <script type="py" src="/src/pyscript/worker.py" worker config="/src/pyconfig.json"></script>
 </body>
 
 </html>
